Move HoC to be with other top-level components
